Added hashing.h My own std::hash specializations will go here
