+++ src/noise2D.cpp
@@ -3,10 +3,8 @@
 #include <zarks/internal/zmath_internals.h>
 #include <zarks/internal/noise_internals.h>
 
-#include <chrono>
 #include <cmath>
-#include <random>
-#include <unordered_map>
+#include <chrono>
 #include <iostream>
 #include <utility>
